Move maxAndMin, dateFormatter to dataParse file
